restructure connect class
