+++ extension/src/connect/inputBox.ts
@@ -1,9 +1,4 @@
-export const isStudioAccessToken = (text?: string): boolean => {
-  if (!text) {
-    return false
-  }
-  return text.startsWith('isat_') && text.length === 54
-}
+import { isStudioAccessToken } from './token'
 
 export const validateTokenInput = (input: string | undefined) => {
   if (!isStudioAccessToken(input)) {
