+++ extension/src/connect/token.ts
@@ -1,0 +1,6 @@
+export const isStudioAccessToken = (text?: string): boolean => {
+  if (!text) {
+    return false
+  }
+  return text.startsWith('isat_') && text.length === 54
+}
